// openmetadata-ui/src/main/resources/ui/src/enums/entity.enum.ts
export enum TabSpecificField {
  SAMPLE_DATA = 'sampleData',
  ACTIVITY_FEED = 'activity_feed',
  TABLE_PROFILE = 'profile',
  LINEAGE = 'lineage',
  COLUMNS = 'columns',
  USAGE_SUMMARY = 'usageSummary',
  FOLLOWERS = 'followers',
  JOINS = 'joins',
  TAGS = 'tags',
  OWNERS = 'owners',
  DATAMODEL = 'dataModel',
  CHARTS = 'charts',
  TASKS = 'tasks',
  TABLE_QUERIES = 'tableQueries',
  TESTS = 'tests',
  PIPELINE_STATUS = 'pipelineStatus',
  PIPELINE_STATUSES = 'pipelineStatuses',
  DASHBOARD = 'dashboard',
  TABLE_CONSTRAINTS = 'tableConstraints',
  EXTENSION = 'extension',
  EXECUTIONS = 'executions',
  SCHEDULE_INTERVAL = 'scheduleInterval',
  TESTSUITE = 'testSuite',
  VIEW_DEFINITION = 'viewDefinition',
  SCHEMA_DEFINITION = 'schemaDefinition',
  FIELDS = 'fields',
  VOTES = 'votes',
  DOMAIN = 'domain',
  DOMAINS = 'domains',
  DATA_PRODUCTS = 'dataProducts',
  REVIEWERS = 'reviewers',
  PIPELINES = 'pipelines',
  PROFILE = 'profile',
  TEAMS = 'teams',
  USERS = 'users',
  ROLES = 'roles',
  PERSONAS = 'personas',
  ASSETS = 'assets',
  EXPERTS = 'experts',
  DEFAULT_PERSONA = 'defaultPersona',
  USAGE_COUNT = 'usageCount',
  PARENT = 'parent',
  PARENTS = 'parents',
  CHILDREN = 'children',
  CUSTOM_METRICS = 'customMetrics',
  TEST_CASE_RESULT = 'testCaseResult',
  INCIDENT_ID = 'incidentId',
  QUERY_USED_IN = 'queryUsedIn',
  TEST_DEFINITION = 'testDefinition',
  SUMMARY = 'summary',
  RELATED_TERMS = 'relatedTerms',
  FOLLOWS = 'follows',
  OWNS = 'owns',
  DATA_INSIGHT_CHART = 'dataInsightChart',
  POLICIES = 'policies',
  DEFAULT_ROLES = 'defaultRoles',
  TERM_COUNT = 'termCount',
  USER_COUNT = 'userCount',
  CHILDREN_COUNT = 'childrenCount',
  START_DATE = 'startDate',
  END_DATE = 'endDate',
  TARGET_DEFINITION = 'targetDefinition',
  TARGET_VALUE = 'targetValue',
  METRIC_TYPE = 'metricType',
  CUSTOM_PROPERTIES = 'customProperties',
  LOCATION = 'location',
  RELATED_METRICS = 'relatedMetrics',
  UI_CUSTOMIZATION = 'uiCustomization',
}
// ...
export enum EntityAction {
  EXPORT = 'export',
  IMPORT = 'import',
}